Simplify slice; factor out benchmark
