+++ src/impls/slice.rs
@@ -1,31 +1,11 @@
-use std::mem;
-use Traversal;
+use {Traversal, Internal};
 
 impl<'a, T> Traversal for &'a [T] {
     type Item = &'a T;
 
     #[inline]
-    fn foreach<F>(self, mut f: F) where F: FnMut(&'a T) -> bool {
-        unsafe {
-            let ptr = self.as_ptr();
-            let len = self.len();
-
-            let is_zero_size = mem::size_of::<T>() == 0;
-
-            if is_zero_size {
-                for _ in 0..len {
-                    // Just give some pointer, doesn't matter what.
-                    if f(mem::transmute(1usize)) { break }
-                }
-            } else {
-                let mut current = ptr;
-                let end = ptr.offset(len as isize);
-                while current != end {
-                    if f(mem::transmute(current)) { break }
-                    current = current.offset(1);
-                }
-            }
-        }
+    fn foreach<F>(self, f: F) where F: FnMut(&'a T) -> bool {
+        Internal::new(self).foreach(f)
     }
 }
 
@@ -33,27 +13,8 @@ impl<'a, T> Traversal for &'a mut [T] {
     type Item = &'a mut T;
 
     #[inline]
-    fn foreach<F>(self, mut f: F) where F: FnMut(&'a mut T) -> bool {
-        unsafe {
-            let ptr = self.as_mut_ptr();
-            let len = self.len();
-
-            let is_zero_size = mem::size_of::<T>() == 0;
-
-            if is_zero_size {
-                for _ in 0..len {
-                    // Just give some pointer, doesn't matter what.
-                    if f(mem::transmute(1usize)) { break }
-                }
-            } else {
-                let mut current = ptr;
-                let end = ptr.offset(len as isize);
-                while current != end {
-                    if f(mem::transmute(current)) { break }
-                    current = current.offset(1);
-                }
-            }
-        }
+    fn foreach<F>(self, f: F) where F: FnMut(&'a mut T) -> bool {
+        Internal::new(self).foreach(f)
     }
 }
 
@@ -73,34 +34,5 @@ mod test {
         let data = [(), (), ()];
         let traversal: Vec<()> = data.map(|&x| x).collect();
         assert_eq!(traversal, data);
-    }
-}
-
-#[cfg(all(test, feature = "nightly"))]
-mod bench {
-
-    use Traversal;
-    use test::Bencher;
-
-    #[bench]
-    fn bench_internal (bench: &mut Bencher) {
-        use rand::random;
-
-        let data: Vec<usize> = (0..10000).map(|_| random()).collect();
-        bench.iter(|| {
-            data.run(|x| { ::test::black_box(x); });
-        });
-    }
-
-    #[bench]
-    fn bench_external (bench: &mut Bencher) {
-        use rand::random;
-
-        let data: Vec<usize> = (0..10000).map(|_| random()).collect();
-        bench.iter(|| {
-            for datum in data.iter() {
-                ::test::black_box(datum);
-            }
-        });
     }
 }
